+++ crates/hale-codegen/src/stdlib/mod.rs
@@ -12,3 +12,4 @@
 pub(crate) mod decimal;
 pub(crate) mod env;
 pub(crate) mod process;
+pub(crate) mod time;
